:tada: feat: rewrite eden
