+++ .gitignore
@@ -3,3 +3,4 @@
 node_modules
 .pnpm-debug.log
 dist
+trace
