+++ src/treaty/utils.ts
@@ -1,15 +1,9 @@
-import type { LegacyEdenCall } from './types'
-
-export const camelToDash = (str: string) =>
-    str.replace(/[A-Z]/g, (letter) => `-${letter.toLowerCase()}`)
-
 export const composePath = (
     domain: string,
     path: string,
-    query: LegacyEdenCall['$query'] | undefined
+    query: Record<string, string> | undefined
 ) => {
     if (!domain.endsWith('/')) domain += '/'
-    path = camelToDash(path.replace(/index/g, ''))
 
     if (!query || !Object.keys(query).length) return `${domain}${path}`
